requirements.txt
flask
requests
python-dotenv
openai
streamlit==1.46.1
whisper
torch
pydub
sounddevice
scipy
